furhter updates from aiida-diff
see https://github.com/aiidateam/aiida-diff/commit/2ae3ac19e894c67cfe861b5f585677ebc02e6fae
