+++ check-syntax.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-cookiecutter --no-input -f . version=1.0.0a1
+cookiecutter --no-input -f . version=1.0.0b1
 pip install -e aiida-diff[docs,pre-commit,testing]
 cd aiida-diff
 git init && git add -A
