+++ {{cookiecutter.plugin_name}}/.gitignore
@@ -9,3 +9,6 @@
 .coverage
 .pytest_cache
 .vscode
+build/
+dist/
+pip-wheel-metadata/
