+++ {{cookiecutter.plugin_name}}/conftest.py
@@ -6,43 +6,55 @@ from __future__ import absolute_import
 import tempfile
 import shutil
 import pytest
-import os
 
 from aiida.manage.fixtures import fixture_manager
-
-def get_backend_str():
-    """ Return database backend string.
-
-    Reads from 'TEST_AIIDA_BACKEND' environment variable.
-    Defaults to django backend.
-    """
-    from aiida.backends.profile import BACKEND_DJANGO, BACKEND_SQLA
-    backend_env = os.environ.get('TEST_AIIDA_BACKEND')
-    if not backend_env: 
-        return BACKEND_DJANGO
-    elif  backend_env in (BACKEND_DJANGO, BACKEND_SQLA):
-        return backend_env
-
-    raise ValueError("Unknown backend '{}' read from TEST_AIIDA_BACKEND environment variable".format(backend_env))
 
 
 @pytest.fixture(scope='session', autouse=True)
 def aiida_profile():
-    """setup a test profile for the duration of the tests"""
+    """Set up a test profile for the duration of the tests"""
     with fixture_manager() as fixture_mgr:
         yield fixture_mgr
 
 
 @pytest.fixture(scope='function', autouse=True)
 def clear_database(aiida_profile):
-    """clear the database after each test"""
+    """Clear the database after each test"""
     yield
     aiida_profile.reset_db()
 
 
 @pytest.fixture(scope='function')
 def new_workdir():
-    """get a new temporary folder to use as the computer's workdir"""
+    """Get a temporary folder to use as the computer's work directory."""
     dirpath = tempfile.mkdtemp()
     yield dirpath
     shutil.rmtree(dirpath)
+
+
+@pytest.fixture(scope='function')
+def aiida_localhost_computer(new_workdir):
+    """Get an AiiDA computer for localhost.
+
+    :return: The computer node
+    :rtype: :py:class:`aiida.orm.Computer`
+    """
+    from {{cookiecutter.module_name}}.helpers import get_computer
+
+    computer = get_computer(workdir=new_workdir)
+
+    return computer
+
+
+@pytest.fixture(scope='function')
+def aiida_code(aiida_localhost_computer):
+    """Get an AiiDA code.
+
+    :return: The code node
+    :rtype: :py:class:`aiida.orm.Code`
+    """
+    from {{cookiecutter.module_name}}.helpers import get_code
+
+    code = get_code(entry_point='{{cookiecutter.entry_point_prefix}}', computer=aiida_localhost_computer)
+
+    return code
