+++ {{cookiecutter.plugin_name}}/pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+filterwarnings =
+    ignore::DeprecationWarning:aiida:
+    ignore::DeprecationWarning:plumpy:
+    ignore::DeprecationWarning:django:
+    ignore::DeprecationWarning:yaml:
