--- {{cookiecutter.plugin_name}}/{{cookiecutter.module_name}}/.gitignore
@@ -1,1 +1,0 @@
-*.pyc
