+++ {{cookiecutter.plugin_name}}/{{cookiecutter.module_name}}/__init__.py
@@ -7,7 +7,3 @@
 from __future__ import absolute_import
 
 __version__ = "{{cookiecutter.version}}"
-
-# disable psycopg2 warning
-import warnings
-warnings.filterwarnings("ignore", category=UserWarning, module='psycopg2')
