+++ {{cookiecutter.plugin_name}}/{{cookiecutter.module_name}}/calculations.py
@@ -27,6 +27,7 @@ class DiffCalculation(CalcJob):
         """Define inputs and outputs of the calculation."""
         # yapf: disable
         super(DiffCalculation, cls).define(spec)
+        spec.input('metadata.options.resources', valid_type=dict, default={'num_machines': 1, 'num_mpiprocs_per_machine': 1})
         spec.input('metadata.options.parser_name', valid_type=six.string_types, default='{{cookiecutter.entry_point_prefix}}')
         spec.input('metadata.options.output_filename', valid_type=six.string_types, default='patch.diff')
         spec.input('parameters', valid_type=DiffParameters, help='Command line parameters for diff')
