+++ {{cookiecutter.plugin_name}}/{{cookiecutter.module_name}}/tests/test_calculations.py
@@ -5,21 +5,14 @@ from __future__ import print_function
 from __future__ import absolute_import
 
 import os
-import {{cookiecutter.module_name}}.tests as tests
-import pytest
+from {{cookiecutter.module_name}} import tests
 
 
-
-@pytest.mark.process_execution
-def test_process(new_workdir):  # pylint: disable=too-many-locals
+def test_process(aiida_code):
     """Test running a calculation
     note this does not test that the expected outputs are created of output parsing"""
     from aiida.plugins import DataFactory, CalculationFactory
     from aiida.engine import run
-
-    # get code
-    computer = tests.get_computer(workdir=new_workdir)
-    code = tests.get_code(entry_point='{{cookiecutter.entry_point_prefix}}', computer=computer)
 
     # Prepare input parameters
     DiffParameters = DataFactory('{{cookiecutter.entry_point_prefix}}')
@@ -32,23 +25,15 @@ def test_process(new_workdir):  # pylint: disable=too-many-locals
         file=os.path.join(tests.TEST_DIR, "input_files", 'file2.txt'))
 
     # set up calculation
-    options = {
-        "resources": {
-            "num_machines": 1,
-            "num_mpiprocs_per_machine": 1,
-        },
-        "max_wallclock_seconds": 30,
-    }
-
     inputs = {
-        'code': code,
+        'code': aiida_code,
         'parameters': parameters,
         'file1': file1,
         'file2': file2,
         'metadata': {
-            'options': options,
-            'label': "{{cookiecutter.module_name}} test",
-            'description': "Test job submission with the {{cookiecutter.module_name}} plugin",
+            'options': {
+                'max_wallclock_seconds': 30
+            },
         },
     }
 
@@ -57,5 +42,3 @@ def test_process(new_workdir):  # pylint: disable=too-many-locals
 
     assert 'content1' in computed_diff
     assert 'content2' in computed_diff
-
-
